Install ejabberd auth script during build
