+++ AppController/djinn.rb
@@ -4785,7 +4785,6 @@ HOSTS
     my_public = my_node.public_ip
     Ejabberd.stop()
     Djinn.log_run("rm -f /var/lib/ejabberd/*")
-    Ejabberd.write_auth_script(my_public, get_db_master.private_ip, @@secret)
     Ejabberd.write_config_file(my_public)
     Ejabberd.start()
   end
