+++ AppController/lib/ejabberd.rb
@@ -84,163 +84,6 @@ module Ejabberd
     }
   end
 
-  def self.write_auth_script(login_ip, uaserver_ip, secret)
-    auth_script = <<SCRIPT
-#!/usr/bin/python
-
-import sys, logging, struct, hashlib, re, SOAPpy, socket, time
-from struct import *
-
-# logging initialization
-
-sys.stderr = open("/var/log/ejabberd/extauth_err.log", 'a')
-logging.basicConfig(level=logging.INFO,
-  format='%(asctime)s %(levelname)s %(message)s',
-  filename='/var/log/ejabberd/extauth.log',
-  filemode='a')
-
-logging.info("extauth script started, waiting for ejabberd requests")
-
-# db initialization
-
-login_ip = "#{login_ip}"
-uaserver_ip = "#{uaserver_ip}"
-secret = "#{secret}"
-
-uaserver_address = "https://" + uaserver_ip + ":4343"
-server = SOAPpy.SOAPProxy(uaserver_address)
-
-# helper functions
-
-class EjabberdInputError(Exception):
-  def __init__(self, value):
-    self.value = value
-  def __str__(self):
-    return repr(self.value)
-
-def ejabberd_in():
-  logging.debug("trying to read 2 bytes from ejabberd:")
-  try:
-    input_length = sys.stdin.read(2)
-  except IOError:
-    logging.debug("ioerror")
-
-  if len(input_length) is not 2:
-    logging.debug("ejabberd sent us wrong things!")
-    raise EjabberdInputError('Wrong input from ejabberd!')
-
-  logging.debug('got 2 bytes via stdin: %s'%input_length)
-  (size,) = unpack('>h', input_length)
-
-  logging.debug('size of data: %i'%size)
-  income = sys.stdin.read(size).split(':')
-
-  logging.debug("incoming data: %s"%income)
-  return income
-
-def ejabberd_out(bool):
-  logging.debug("Ejabberd gets: %s" % bool)
-  token = genanswer(bool)
-  logging.debug("sent bytes: %#x %#x %#x %#x" % (ord(token[0]), ord(token[1]), ord(token[2]), ord(token[3])))
-  sys.stdout.write(token)
-  sys.stdout.flush()
-
-def genanswer(bool):
-  answer = 0
-  if bool:
-    answer = 1
-
-  token = pack('>hh', 2, answer)
-  return token
-
-def isuser(in_user, in_host):
-  return True
-  #username = in_user + "@" + in_host
-  #userdata = server.get_user_data(username, secret)
-
-  #if userdata == 'Error: Bad length of user schema vs user result'
-  #  return True
-  #else:
-  #  return False
-
-def auth(in_user, in_host, password):
-  #return True
-  username = in_user + "@" + in_host
-
-  logging.info("trying to authenticate user [%s]" % (username))
-
-  if not isuser(in_user, in_host):
-    return False
-
-  userdata = ""
-  while True:
-    try:
-      userdata = server.get_user_data(username, secret)
-      break
-    except socket.error:
-      time.sleep(1)
-
-  logging.info("userdata for [%s] is [%s]" % (username, str(userdata)))
-  matchdata = re.search('password:(.*)', userdata)
-
-  if matchdata is None:
-    logging.info("matchdata for [%s] was none" % (username))
-    return False
-
-  remote_password = matchdata.group(1)
-
-  salted = username + password
-  local_password = hashlib.sha1(salted).hexdigest()
-
-  logging.info("local  password: [%s]" % (local_password))
-  logging.info("remote password: [%s]" % (remote_password))
-
-  if local_password == remote_password:
-    return True
-  else:
-    return False
-
-def log_result(op, in_user, bool):
-  if bool:
-    logging.info("%s successful for %s"%(op, in_user))
-  else:
-    logging.info("%s unsuccessful for %s"%(op, in_user))
-
-# main loop
-
-while True:
-  logging.info("start of infinite loop")
-  try: 
-    ejab_request = ejabberd_in()
-  except EjabberdInputError, inst:
-    logging.info("Exception occurred: %s", inst)
-    break
-
-  logging.debug('operation: %s'%(ejab_request[0]))
-  op_result = False
-
-  if ejab_request[0] == "auth":
-    op_result = auth(ejab_request[1], ejab_request[2], ejab_request[3])
-    ejabberd_out(op_result)
-    log_result(ejab_request[0], ejab_request[1], op_result)
-  elif ejab_request[0] == "isuser":
-    op_result = isuser(ejab_request[1], ejab_request[2])
-    ejabberd_out(op_result)
-    log_result(ejab_request[0], ejab_request[1], op_result)
-  elif ejab_request[0] == "setpass":
-    op_result=False
-    ejabberd_out(op_result)
-    log_result(ejab_request[0], ejab_request[1], op_result)
-
-logging.info("extauth script terminating")
-
-SCRIPT
-
-    HelperFunctions.write_file(AUTH_SCRIPT_LOCATION, auth_script)
-    Djinn.log_run("chown ejabberd #{AUTH_SCRIPT_LOCATION}")
-    Djinn.log_run("chmod +x #{AUTH_SCRIPT_LOCATION}")
-  end
-
   def self.write_config_file(my_private_ip)
     config = <<CONFIG
 %%%
